Fabrice Kauffmann:
This is my latest model: the F-18 Hornet from the VFC-13 "Top Gun Aggressors".
The camouflage imitates a Russian Sukhoï-27 Flanker. An alternate version of
the textures is available under the Models/desert directory.
Additional textures should be available in the future, as well as a 3D cockpit.

Screenshots: http://scifi.skynetblogs.be/
Download URL: http://fkauffmann.ifrance.com/archives/