Fabrice Kauffmann:
Please find below a new update (0.5) for my F-18 Hornet.
I was not very satisfied with the first release of the 3D cockpit (too much
bitmaps).  Therefore, I decided to remove all those flat textures and to
add a maximum of 3d objects.

As usually, you can see pictures on my blog:

    http://scifi.skynetblogs.be/

The model is available for download at the following URL:

    http://fkauffmann.ifrance.com/archives/